requirements.txt
python-telegram-bot==21.11.1
aiohttp==3.9.1
